% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{create_mentions_edgelist}
\alias{create_mentions_edgelist}
\title{Create an edgelist where interaction is defined by mentioning}
\usage{
create_mentions_edgelist(df)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}
}
\value{
A dataframe edgelist with column names 'sender', 'receiver', and
  \code{edge_type}, which in this case the edge type is "mention"
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{create_mentions_edgelist()} pulls out senders and receivers of
  mentions and adds a new column \code{edge_type}.
  \code{create_mentions_edgelist()} is a useful function in itself, but is
  also used in \code{create_edgelist()}.
}
\examples{

\dontrun{

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url))
mentions_edgelist <- create_mentions_edgelist(tmp_df)
mentions_edgelist
}
}
\seealso{
Compare to other \code{tidytags} functions such as
  \code{create_replies_edgelist()}, \code{create_retweets_edgelist()},
  \code{create_quotes_edgelist()}, and \code{create_edgelist()}.
}
