% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{filter_by_tweet_type}
\alias{filter_by_tweet_type}
\title{Create an edgelist where interaction is defined by replying}
\usage{
filter_by_tweet_type(df, type)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}

\item{type}{The specific kind of tweets that will be kept in the dataset
after filtering the rest. Choices for \code{type}include"reply", "retweet",
"quote", or "mention." Defaults to "all."}
}
\value{
A dataframe of processed tweets and fewer rows. Only the tweets of
  the specified type will remain.
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{create_edgelist()} processes the tweets by calling
  \code{process_tweets()} and filters the dataset to only keep the specific
  tweet types (e.g., replies, retweets, quote tweets, and mentions)
  requested. \code{filter_by_tweet_type()} is a useful function in itself,
  but it is also used in \code{create_edgelist()}.
}
\examples{

\dontrun{

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url))
tmp_re <- filter_by_tweet_type(tmp_df, "reply")
tmp_rt <- filter_by_tweet_type(tmp_df, "retweet")
tmp_qu <- filter_by_tweet_type(tmp_df, "quote")
tmp_me <- filter_by_tweet_type(tmp_df, "mention")
}
}
