% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{get_mentions}
\alias{get_mentions}
\title{Create an edgelist where interaction is defined by mentioning}
\usage{
get_mentions(df)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}
}
\value{
A dataframe edgelist with column names 'sender', 'receiver', and
  \code{edge_type}, which in this case the edge type is "mention"
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{get_mentions()} pulls out senders and receivers of mentions and adds
  a new column \code{edge_type}. \code{get_mentions()} is a useful function
  in itself, but is also used in \code{create_edgelist()}.
}
\seealso{
Compare to other \code{rTags} functions such as \code{get_replies()},
  \code{get_retweets()}, \code{get_quotes()}, and \code{create_edgelist()}.
}
