% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{get_quotes}
\alias{get_quotes}
\title{Create an edgelist where interaction is defined by quote-tweeting}
\usage{
get_quotes(df)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}
}
\value{
A dataframe edgelist with column names 'sender', 'receiver', and
  \code{edge_type}, which in this case the edge type is "quote-tweet"
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{get_quotes()} pulls out senders and receivers of quote tweets and adds
  a new column \code{edge_type}. \code{get_quotes()} is a useful function
  in itself, but is also used in \code{create_edgelist()}.
}
\seealso{
Compare to other \code{tidtags} functions such as \code{get_replies()},
  \code{get_retweets()}, \code{get_mentions()}, and \code{create_edgelist()}.
}
