% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{get_replies}
\alias{get_replies}
\title{Create an edgelist where interaction is defined by replying}
\usage{
get_replies(df)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}
}
\value{
A dataframe edgelist with column names 'sender', 'receiver', and
  \code{edge_type}, which in this case the edge type is "reply"
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{get_replies()} pulls out senders and receivers of reply tweets and
  adds a new column \code{edge_type}. \code{get_replies()} is a useful function
  in itself, but is also used in \code{create_edgelist()}.
}
\seealso{
Compare to other \code{tidtags} functions such as \code{get_retweets()},
  \code{get_quotes()}, \code{get_mentions()}, and \code{create_edgelist()}.
}
