% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{get_retweets}
\alias{get_retweets}
\title{Create an edgelist where interaction is defined by retweeting}
\usage{
get_retweets(df)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}
}
\value{
A dataframe edgelist with column names 'sender', 'receiver', and
  \code{edge_type}, which in this case the edge type is "retweet"
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{get_retweets()} pulls out senders and receivers of retweets and adds
  a new column \code{edge_type}. \code{get_retweets()} is a useful function
  in itself, but is also used in \code{create_edgelist()}.
}
\examples{

\dontrun{

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
get_retweets(tmp_df)
}
}
\seealso{
Compare to other \code{tidytags} functions such as \code{get_replies()},
  \code{get_quotes()}, \code{get_mentions()}, and \code{create_edgelist()}.
}
