% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-upstream-tweets.R
\name{get_upstream_tweets}
\alias{get_upstream_tweets}
\title{Collect upstream tweets and add to dataset}
\usage{
get_upstream_tweets(df)
}
\arguments{
\item{df}{A dataframe of tweets and full metadata from the Twitter API as
returned by \code{pull_tweet_data()}}
}
\value{
A new, expanded dataframe which includes any retrievable upstream
  replies
}
\description{
Because the Twitter API offers a \code{reply_to_status_id} column, it is
  possible to iteratively reconstruct reply threads in an *upstream*
  direction, that is, retrieving tweets composed earlier than replies in the
  dataset. The \code{get_upstream_tweets()} function collects upstream
  replies not previously found in the dataset. Keep in mind that there is no
  way to predict how far upstream you can trace back a reply thread, so
  running \code{get_upstream_tweets()} could take a while and potentially hit
  the Twitter API rate limit of 90,000 tweets in a 15-minute period.
}
\details{
This function requires authentication; please see
  \code{vignette("setup", package = "tidytags")}
}
\examples{

\dontrun{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url))
more_replies_df <- get_upstream_tweets(tmp_df)
more_replies_df
}
}
