% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{length_with_na}
\alias{length_with_na}
\title{Count the number of items in a list, with an NA entry returning 0}
\usage{
length_with_na(x)
}
\arguments{
\item{x}{A list}
}
\value{
The number of items in the list
}
\description{
Count the number of items in a list, with an NA entry returning 0
}
\keyword{internal}
