% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{lookup_many_tweets}
\alias{lookup_many_tweets}
\title{Retrieve the fullest extent of metadata for more than 90,000 statuses}
\usage{
lookup_many_tweets(x, alarm = FALSE)
}
\arguments{
\item{x}{A list or vector of status ID numbers}

\item{alarm}{An audible notification that a batch of 90,000 statuses has been
completed}
}
\value{
A dataframe of statuses and full metadata from the Twitter API
}
\description{
This function calls \code{pull_tweet_data()}, but has a built-in delay
of 15 minutes to allow the Twitter API to reset after looking up 90,000
statuses
}
\details{
This function requires authentication; please see
\code{vignette("setup", package = "tidytags")}
}
\examples{
\dontrun{

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)
lookup_many_tweets(tags_content$id_str)
lookup_many_tweets("1176592704647716864")
lookup_many_tweets("1176592704647716864", alarm = TRUE)
}
}
\seealso{
Read more about rtweet authentication setup at
\code{vignette("auth", package = "rtweet")}
}
