% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{lookup_many_users}
\alias{lookup_many_users}
\title{Retrieve the fullest extent of tweet metadata for more than 90,000 users}
\usage{
lookup_many_users(x, alarm = FALSE)
}
\arguments{
\item{x}{A list or vector of user ID numbers}

\item{alarm}{An audible notification that a batch of 90,000 users has been
completed}
}
\value{
A dataframe of tweets and full user metadata from the Twitter API
}
\description{
This function calls \code{rtweet::lookup_users()}, but has a built-in delay
of 15 minutes to allow the Twitter API to reset after looking up 90,000
users.
}
\details{
This function requires authentication; please see
\code{vignette("setup", package = "tidytags")}
}
\examples{
\dontrun{

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url))
users <- lookup_many_users(tmp_df$screen_name, alarm = TRUE)
users

lookup_many_users("AECT")
lookup_many_tweets("12030342", alarm = TRUE)
}
}
\seealso{
Read more about \code{library(rtweet)}
\href{https://rtweet.info/}{here}.
}
