% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{process_tweets_flattened}
\alias{process_tweets_flattened}
\title{Calculate additional information using tweet metadata}
\usage{
process_tweets_flattened(df)
}
\arguments{
\item{df}{A dataframe of tweets and full metadata from the Twitter API}
}
\value{
A dataframe with several additional columns: word_count, character_count,
  mentions_count, hashtags_count_api, hashtags_count_regex, has_hashtags,
  urls_count_api, urls_count_regex, is_reply, is_self_reply
}
\description{
Calculate additional information using tweet metadata
}
