% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{read_tags}
\alias{read_tags}
\title{Retrieve a TAGS archive of tweets and bring into R}
\usage{
read_tags(url)
}
\arguments{
\item{url}{A valid URL (i.e., hyperlink) to a TAGS tracker}
}
\value{
A dataframe of the TAGS archive of tweets
}
\description{
Retrieve a TAGS archive of tweets and bring into R
}
\seealso{
Read more about \code{library(googlesheets)} \href{https://github.com/jennybc/googlesheets}{here}.
}
