% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-users-data.R
\name{add_users_data}
\alias{add_users_data}
\title{Retrieve user information for everyone in an edgelist}
\usage{
add_users_data(edgelist, users_data_from_rtweet)
}
\arguments{
\item{edgelist}{An edgelist of senders and receivers, such as that returned by
the function \code{create_edgelist()}.}

\item{users_data_from_rtweet}{A dataframe with Twitter users data retrieved using
the function \code{rtweet::users_data()}.}
}
\value{
A dataframe in the form of an edgelist (i.e., with senders and receivers)
  as well as numerous, appropriately named columns of details about the senders
  and receivers.
}
\description{
\code{add_users_data()} updates an edgelist created with \code{create_edgelist()}
  by appending user data retrieved with \code{rtweet::users_data()}. The resulting
  dataframe keeps all the columns from \code{rtweet} but adds "_sender" or
  "_receiver" to the column names
}
\seealso{
Review documentation for \code{rtweet::users_data()} for a full list
  of metadat retrieved (i.e., column names) by this function.
}
