% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-users-data.R
\name{add_users_data}
\alias{add_users_data}
\title{Retrieve user information for everyone in an edgelist}
\usage{
add_users_data(edgelist)
}
\arguments{
\item{edgelist}{An edgelist of senders and receivers, such as that returned
by the function \code{create_edgelist()}.}
}
\value{
A dataframe in the form of an edgelist (i.e., with senders and
receivers) as well as numerous, appropriately named columns of details
about the senders and receivers.
}
\description{
Updates an edgelist created with \code{create_edgelist()} by appending user
data retrieved with \code{rtweet::lookup_users()}. The resulting dataframe
adds many additional columns and appends "_sender" or "_receiver" to the
column names.
}
\details{
This function requires authentication; please see
\code{vignette("setup", package = "tidytags")}
}
\examples{

\donttest{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)

if (rtweet::auth_has_default()) {
  tweets_data <- lookup_many_tweets(tags_content)
  add_users_data(create_edgelist(tweets_data))
}
}

}
\seealso{
Read more about rtweet authentication setup at
\code{vignette("auth", package = "rtweet")}
}
