% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{create_edgelist}
\alias{create_edgelist}
\title{Create an edgelist with several types of interaction}
\usage{
create_edgelist(df)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}
}
\value{
A dataframe edgelist with column names 'sender', 'receiver', and
  \code{edge_type}, which in this case the edge type may be "reply",
  "retweet", "quote", or "mention"
}
\description{
Starting with a dataframe returned by \code{pull_tweet_data()},
  \code{create_edgelist()} pulls out senders and receivers of reply tweets,
  retweets, quote tweets, and mentions, and then adds a new column
  \code{edge_type}.
}
\examples{

\dontrun{

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
edgelist <- create_edgelist(tmp_df)
edgelist
}
}
\seealso{
Compare to other \code{tidytags} functions such as
  \code{create_replies_edgelist()}, \code{create_retweets_edgelist()},
  \code{create_quotes_edgelist()}, and \code{create_mentions_edgelist()}.
}
