% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{create_edgelist}
\alias{create_edgelist}
\title{Create an edgelist where senders and receivers are defined by different types
of Twitter interactions}
\usage{
create_edgelist(df, type = "all")
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}

\item{type}{The specific kind of statuses used to define the interactions
around which the edgelist will be built. Choices include "reply",
"retweet", or "quote". Defaults to "all".}
}
\value{
A dataframe edgelist defined by interactions through the type of
statuses specified. The dataframe has three columns: \code{sender},
\code{receiver}, and \code{edge_type}.
}
\description{
Starting with a dataframe of Twitter data imported to R with
\code{read_tags()} and additional metadata retrieved by
\code{pull_tweet_data()}, \code{create_edgelist()} removes any statuses
that are not of the requested type (e.g., replies, retweets, and quote
tweets) by calling \code{filter_by_tweet_type()}. Finally, \code{create_edgelist()}
pulls out senders and receivers of the specified type of statuses, and then
adds a new column called \code{edge_type}.
}
\examples{

\donttest{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)

if (rtweet::auth_has_default()) {
  tweets_data <- lookup_many_tweets(tags_content)
  full_edgelist <- create_edgelist(tweets_data)
  full_edgelist

  reply_edgelist <- create_edgelist(tweets_data, type = "reply")
  retweet_edgelist <- create_edgelist(tweets_data, type = "retweet")
  quote_edgelist <- create_edgelist(tweets_data, type = "quote")
}
}

}
