% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-network.R
\name{filter_by_tweet_type}
\alias{filter_by_tweet_type}
\title{Filter a Twitter dataset to only include statuses of a particular type}
\usage{
filter_by_tweet_type(df, type)
}
\arguments{
\item{df}{A dataframe returned by \code{pull_tweet_data()}}

\item{type}{The specific kind of statuses that will be kept in the dataset
after filtering the rest. Choices for \code{type}include "reply",
"retweet", "quote", and "original".}
}
\value{
A dataframe of processed statuses and fewer rows that the input
dataframe. Only the statuses of the specified type will remain.
}
\description{
Starting with a dataframe of Twitter data imported to R with
\code{read_tags()} and additional metadata retrieved by
\code{pull_tweet_data()}, \code{filter_by_tweet_type()} processes the
statuses by calling \code{process_tweets()} and then removes any statuses
that are not of the requested type (e.g., replies, retweets, and quote
tweets). \code{filter_by_tweet_type()} is a useful function in itself, but it is
also used in \code{create_edgelist()}.
}
\examples{

\donttest{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)

if (rtweet::auth_has_default()) {
  tweets_data <- lookup_many_tweets(tags_content)
  only_replies <- filter_by_tweet_type(tweets_data, "reply")
  only_retweets <- filter_by_tweet_type(tweets_data, "retweet")
  only_quote_tweets <- filter_by_tweet_type(tweets_data, "quote")
  only_originals <- filter_by_tweet_type(tweets_data, "original")
}
}

}
