% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-geography.R
\name{geocode_tags}
\alias{geocode_tags}
\title{Retrieve geographic coordinates}
\usage{
geocode_tags(df, google_key = Sys.getenv("Google_API_key"))
}
\arguments{
\item{df}{A dataframe or tibble}

\item{google_key}{An Google Developers API key which will need to be obtained
by the user}
}
\value{
Geographic coordinates (i.e., latitude and longitude) which may then be
  used to plot locations on a map
}
\description{
\code{geocode_tags} retrieves geographic coordinates (i.e., latitude and
  longitude) based on the locations listed in Twitter user profiles.
}
\seealso{
A key starting place is Google's
  \href{https://developers.google.com/maps/documentation/geocoding/intro}{
  geocoding documentation} for developers. Blog posts from
  \href{https://www.jessesadler.com/post/geocoding-with-r/}{Jesse Sadler}
  and \href{https://www.littlemissdata.com/blog/maps}{Laura Ellis}
  may also provide additional inspiration for geocoding. Finally, the
  \code{ggmap} package can provide additional functionality for visualizing
  geographic data.
}
