% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{get_char_tweet_ids}
\alias{get_char_tweet_ids}
\title{Get Twitter status ID numbers as character strings}
\usage{
get_char_tweet_ids(x)
}
\arguments{
\item{x}{A dataframe containing the column name 'status_url'
(i.e., the hyperlink to specific statuses), such as that returned by
\code{read_tags()}, or a vector of status URLs, such as as those contained
in the 'status_url' column of a dataframe returned by
\code{tidytags::read_tags()}}
}
\value{
A vector of Twitter status IDs as character strings
}
\description{
This function is useful because Google Sheets (and hence TAGS)
typically round very large numbers into an exponential form. Thus,
because status ID numbers are very large, they often get corrupted in this
rounding process. The most reliable way to get full status ID numbers is by
using this function, \code{get_char_tweet_ids()}, to pull the ID numbers
from the URL linking to specific statuses.
}
\examples{

\donttest{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)
get_char_tweet_ids(tags_content[1:10, ])
get_char_tweet_ids(tags_content$status_url[1:10])
get_char_tweet_ids(
  "https://twitter.com/tweet__example/status/1176592704647716864")
}

}
