% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-upstream-replies.R
\name{get_upstream_replies}
\alias{get_upstream_replies}
\title{Collect upstream replies and add to dataset}
\usage{
get_upstream_replies(df)
}
\arguments{
\item{df}{A dataframe of tweets and full metadata from the Twitter API as
returned by \code{pull_tweet_data()}}
}
\value{
A new, expanded dataframe which includes any retrievable upstream
  replies
}
\description{
\code{get_upstream_replies()} collects upstream replies not previously found
  in the dataset.
}
\examples{

\dontrun{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
new_df <- get_upstream_replies(tmp_df)
new_df
}
}
