% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-upstream-replies.R
\name{get_upstream_replies}
\alias{get_upstream_replies}
\title{Collect upstream replies and add to dataset}
\usage{
get_upstream_replies(df)
}
\arguments{
\item{df}{A dataframe or tibble}
}
\value{
A new dataframe which includes any retrievable upstream replies
}
\description{
\code{get_upstream_replies()} collects upstream replies not previously found
  in the dataset.
}
