% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-upstream-tweets.R
\name{get_upstream_tweets}
\alias{get_upstream_tweets}
\title{Collect upstream statuses and add to dataset}
\usage{
get_upstream_tweets(df)
}
\arguments{
\item{df}{A dataframe of statuses and full metadata from the Twitter API as
returned by \code{pull_tweet_data()}}
}
\value{
A new, expanded dataframe which includes any retrievable upstream
replies
}
\description{
Because the Twitter API offers a \code{in_reply_to_status_id_str} column, it is
possible to iteratively reconstruct reply threads in an \emph{upstream}
direction, that is, retrieving statuses composed earlier than replies in
the dataset. The \code{get_upstream_tweets()} function collects upstream
replies not previously found in the dataset. Keep in mind that there is no
way to predict how far upstream you can trace back a reply thread, so
running \code{get_upstream_tweets()} could take a while and potentially hit
the Twitter API rate limit of 90,000 statuses in a 15-minute period.
}
\details{
This function requires authentication; please see
\code{vignette("setup", package = "tidytags")}
}
\examples{

\donttest{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)

if (rtweet::auth_has_default()) {
  tweets_data <- lookup_many_tweets(tags_content)
  more_replies_df <- get_upstream_tweets(tweets_data)
  more_replies_df
}
}

}
\seealso{
Read more about rtweet authentication setup at
\code{vignette("auth", package = "rtweet")}
}
