% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{process_tweets}
\alias{process_tweets}
\title{Calculate additional information using status metadata}
\usage{
process_tweets(df)
}
\arguments{
\item{df}{A dataframe of statuses and full metadata from the Twitter API as
returned by \code{pull_tweet_data()}}
}
\value{
A dataframe with several additional columns: mentions_count,
hashtags_count, urls_count, tweet_type, is_self_reply
}
\description{
Calculate additional information using status metadata
}
\examples{

\donttest{
example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)

if (rtweet::auth_has_default()) {
  tweets_data <- lookup_many_tweets(tags_content)
  tweets_processed <- process_tweets(tweets_data)
  tweets_processed
}
}

}
