% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{pull_tweet_data}
\alias{pull_tweet_data}
\title{Retrieve the fullest extent of tweet metadata available from the Twitter API}
\usage{
pull_tweet_data(x, n = NULL)
}
\arguments{
\item{x}{A list or vector of tweet ID numbers}

\item{n}{The number of tweets to look up, by default the total number of tweet
ID numbers available, but capped at 90,000 due to Twitter API limitations.}
}
\value{
A dataframe of tweets and full metadata from the Twitter API
}
\description{
With TAGS imported into R, `pull_tweet_data()` uses the **rtweet** package
  to query the Twitter API. Using rtweet requires a Twitter developer account;
  see the rtweet vignette \href{https://rtweet.info/articles/auth.html}{Obtaining and using access tokens}
  as a guide to get started.
}
\seealso{
Read more about \code{library(rtweet)} \href{https://rtweet.info/}{here}.
}
