% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-functions.R
\name{read_tags}
\alias{read_tags}
\title{Retrieve a TAGS archive of tweets and bring into R}
\usage{
read_tags(url)
}
\arguments{
\item{url}{A valid URL (i.e., hyperlink) to a TAGS tracker}
}
\value{
A dataframe of the TAGS archive of tweets
}
\description{
Keep in mind that `read_tags()` uses the **googlesheets4** package,
  and one requirement is that your TAGS has been published to the web.
  To do this, with the TAGS page open in a browser, go to
  `File` >> `Publish to the web`. The `Link` field should be 'Entire document'
  and the `Embed` field should be 'Web page.' If everything looks right,
  then click the `Publish` button. Next, click the `Share` button in the
  top right corner of the Google Sheets window, select `Get shareable link`,
  and set the permissions to 'Anyone with the link can view.' Now you should
  be ready to go.
}
\examples{
  \dontrun{
  example_url <- "https://docs.google.com/spreadsheets/d/18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8/edit#gid=8743918"
  read_tags(example_url)
  }
}
\seealso{
Read more about \code{library(googlesheets4)} \href{https://github.com/tidyverse/googlesheets4}{here}.
}
